Django-projects/TODOIT_API/todoitapiproject/todoitapiproject/urls.py
"""todoitapiproject URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/4.1/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path, include, re_path
from django.conf import settings
from django.conf.urls.static import static
from rest_framework.authtoken import views
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
    TokenVerifyView
)
from accounts import views as accounts_views
from todoitapiproject.docs import urlpatterns as docs_urlpatterns


urlpatterns = [
    path('admin/', admin.site.urls),
    # path('api-auth/', include('rest_framework.urls')),
    re_path(r'^api/demo/', include('demoapi.urls')),
    re_path(r'^api/v1/', include('api_v1.urls')),

    # tokens
    path('api-token-auth/', views.obtain_auth_token),
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/token/verify/', TokenVerifyView.as_view(), name='token_verify'),
    path('account-token-auth', accounts_views.CustomAuthToken.as_view()),
    path('account-custom-token', accounts_views.MyObtainTokenPairView.as_view()),
] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)


# docs
urlpatterns += docs_urlpatterns